vm: fix some size_t/uint64_t issues (31)

+++ src/cdisasm/disasm.cpp
@@ -24,7 +24,7 @@ struct StreamReader : public BaseReader
         auto offset     = position();
         auto max_length = std::min(size, length - offset);
 
-        memcpy(buffer, data + offset, max_length);
+        memcpy(buffer, data + offset, (size_t)max_length);
 
         return max_length;
     }
